+++ libs/translator/reader/read_geometry.cpp
@@ -35,12 +35,9 @@
 #include <pxr/base/gf/rotation.h>
 #include <pxr/base/gf/transform.h>
 #include <pxr/base/tf/stringUtils.h>
-
-#include <ai.h>
-#include <cstdio>
-#include <cstring>
-#include <string>
-#include <vector>
+#include <pxr/usd/usdLux/nonboundableLightBase.h>
+#include <pxr/usd/usdLux/boundableLightBase.h>
+#include <pxr/usd/usdLux/lightAPI.h>
 
 #include <constant_strings.h>
 #include <shape_utils.h>
@@ -55,6 +52,7 @@ PXR_NAMESPACE_USING_DIRECTIVE
 // clang-format off
 TF_DEFINE_PRIVATE_TOKENS(
     _tokens,
+    (LightAPI)
     ((PrimvarsArnoldLightShaders, "primvars:arnold:light:shaders"))
 );
 
@@ -1377,12 +1375,40 @@ AtNode* UsdArnoldReadPointInstancer::Read(const UsdPrim &prim, UsdArnoldReaderCo
     // as they need to treat instance matrices differently
     std::vector<bool> nodesChildProcs(protoPaths.size(), false);    
     int numChildProc = 0;    
+    std::vector<float> protoLightIntensities;
 
     for (size_t i = 0; i < protoPaths.size(); ++i) {
         const SdfPath &protoPath = protoPaths.at(i);
         // get the proto primitive, and ensure it's properly exported to arnold,
         // since we don't control the order in which nodes are read.
         UsdPrim protoPrim = reader->GetStage()->GetPrimAtPath(protoPath);
+
+        // If some of the prototypes are lights we'll need to set a user data
+        // instance_intensity, as we do for meshes visibility. 
+        // There are currently different ways of having light primitives in USD.
+        // Typed schemas can derive from UsdLuxBoundableLightBase or UsdLuxNonboundableLightBase.
+        // But the LightAPI schema can also be applied to any primitive
+        if (protoPrim.IsA<UsdLuxBoundableLightBase>() || 
+            protoPrim.IsA<UsdLuxNonboundableLightBase>() 
+#if PXR_VERSION >= 2302
+            || protoPrim.HasAPI(_tokens->LightAPI)
+#endif
+            ) {
+            
+            // This prototype is a light, let's initialize our 
+            // vector to a default intensity of 1
+            if (protoLightIntensities.empty())
+                protoLightIntensities.assign(protoPaths.size(), 1.f);
+
+            // Get the intensity value from the light primitive
+            float lightIntensity = 1.f;
+            UsdAttribute intensityAttr = protoPrim.GetAttribute(str::t_inputs_intensity);
+            VtValue intensityValue;
+            if (intensityAttr && intensityAttr.Get(&intensityValue, frame)) {
+                protoLightIntensities[i] = VtValueGetFloat(intensityValue);
+            }
+        }
+
         std::string objType = (protoPrim) ? protoPrim.GetTypeName().GetText() : "";
 
         if (protoPrim)
@@ -1471,6 +1497,10 @@ AtNode* UsdArnoldReadPointInstancer::Read(const UsdPrim &prim, UsdArnoldReaderCo
     std::vector<unsigned char> instanceVisibilities(numInstances, AI_RAY_ALL);
     std::vector<unsigned int> instanceIdxs(numInstances, 0);
 
+    std::vector<float> instanceIntensities;
+    if (!protoLightIntensities.empty())
+        instanceIntensities.assign(numInstances, 1.f);
+
     // Create a big matrix array with all the instance matrices for the first key, 
     // then all matrices for the second key, etc..
     std::vector<AtMatrix> instance_matrices(numKeys * numInstances);
@@ -1479,8 +1509,11 @@ AtNode* UsdArnoldReadPointInstancer::Read(const UsdPrim &prim, UsdArnoldReaderCo
         if ((!pruneMaskValues.empty() && pruneMaskValues[i] == false) || 
             (protoIndices[i] >= (int) protoVisibility.size()))
             instanceVisibilities[i] = 0;
-        else
+        else {
             instanceVisibilities[i] = protoVisibility[protoIndices[i]];
+            if (!protoLightIntensities.empty())
+                instanceIntensities[i] = protoLightIntensities[protoIndices[i]];
+        }
 
         // loop over all the motion steps and append the matrices as a big list of floats
         for (size_t t = 0; t < numKeys; ++t) {
@@ -1499,6 +1532,13 @@ AtNode* UsdArnoldReadPointInstancer::Read(const UsdPrim &prim, UsdArnoldReaderCo
     AiNodeSetArray(node, str::instance_matrix, AiArrayConvert(numInstances, numKeys, AI_TYPE_MATRIX, &instance_matrices[0]));
     AiNodeSetArray(node, str::instance_visibility, AiArrayConvert(numInstances, 1, AI_TYPE_BYTE, &instanceVisibilities[0]));
     AiNodeSetArray(node, str::node_idxs, AiArrayConvert(numInstances, 1, AI_TYPE_UINT, &instanceIdxs[0]));
+
+    // If some of the prototypes are lights, we need to set the instance_intensity user data
+    // because the source prototype has its intensity set to 0
+    if (!instanceIntensities.empty()) {
+        AiNodeDeclare(node, str::instance_intensity, "constant ARRAY FLOAT");
+        AiNodeSetArray(node, str::instance_intensity, AiArrayConvert(numInstances, 1, AI_TYPE_FLOAT, &instanceIntensities[0]));
+    }
 
     ReadMatrix(prim, node, time, context);
     InstancerPrimvarsRemapper primvarsRemapper;